requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
audioread==3.0.1
av==12.0.0
basicsr==1.4.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
#dlib==19.24.4
exceptiongroup==1.2.0
face-alignment==1.3.5
facexlib==0.3.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
filterpy==1.4.5
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gfpgan==1.3.8
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.8
librosa==0.9.2
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.9.0.80
orjson==3.10.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
resampy==0.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.7
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
sympy==1.12
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2023.7.10
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
Werkzeug==3.0.2
yacs==0.1.8
yapf==0.40.2
zipp==3.18.1
